+++ Proiect/PaymentMethod.java
@@ -2,7 +2,7 @@ public abstract class PaymentMethod {
     public String getPaymentMethod(){
         return "Unknown";
     }
-    public boolean checkPaymentMethod(int price){
+    public boolean checkPaymentMethod(double price){
         return false;
     }
 }
